pyproject.toml
[tool.poetry]
name = "fractalizer"
version = "1.0.0"
description = "A fractal generator that creates beautiful Mandelbrot, Julia, Phoenix, and Burning Ship fractals with customizable color palettes"
authors = ["devilishclover"]
readme = "readme.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8.1"
colour = "^0.1.5"
pygame = "^2.5.0"

[tool.poetry.group.animation]
optional = true

[tool.poetry.group.animation.dependencies]
opencv-python = "^4.8.0"
numpy = "^1.24.0"
pillow = "^10.0.0"
imageio = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]